RavenDB-17771 moved to processor
